requirements.txt
asgiref==3.9.1
azlyrics==1.3.2
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distlib==0.4.0
Django==5.2.6
django-environ==0.12.0
django-soft-delete==1.0.21
djangorestframework==3.16.1
filelock==3.19.1
Flask==3.1.1
fonttools==4.58.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
nba_api==1.9.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.4.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
virtualenv==20.34.0
Werkzeug==3.1.3
